backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "langchain-mcp-adapters>=0.1.14",
    "langchain[groq]>=1.1.0",
    "langgraph>=1.0.3",
    "pymongo>=4.15.4",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
